ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
langchain==0.1.20
langchain-community==0.0.38
chromadb==0.5.0
sentence-transformers==2.7.0
pymongo==4.7.1
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
apscheduler==3.10.4
lxml==5.2.1
redis==5.0.4
groq==0.9.0
loguru==0.7.2
pydantic==2.7.1
httpx==0.27.0
numpy==1.26.4